pyproject.toml
[project]
name = "gridworldaidemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pydub>=0.25.1",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.1",
    "simpleaudio>=1.0.4",
    "yt-dlp>=2025.8.20",
]
